+++ src/main/java/io/github/hapjava/characteristics/impl/filtermaintenance/FilterChangeIndicationCharacteristic.java
@@ -1,5 +1,6 @@
-package io.github.hapjava.characteristics.impl.occupancysensor;
+package io.github.hapjava.characteristics.impl.filtermaintenance;
 
+import io.github.hapjava.characteristics.EventableCharacteristic;
 import io.github.hapjava.characteristics.HomekitCharacteristicChangeCallback;
 import io.github.hapjava.characteristics.impl.base.EnumCharacteristic;
 import java.util.Optional;
@@ -7,17 +8,17 @@ import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
-/** This characteristic indicates if occupancy was detected (e.g. a person present). */
-public class OccupancyDetectedCharacteristic extends EnumCharacteristic<OccupancyDetectedEnum> {
+public class FilterChangeIndicationCharacteristic
+    extends EnumCharacteristic<FilterChangeIndicationEnum> implements EventableCharacteristic {
 
-  public OccupancyDetectedCharacteristic(
-      Supplier<CompletableFuture<OccupancyDetectedEnum>> getter,
+  public FilterChangeIndicationCharacteristic(
+      Supplier<CompletableFuture<FilterChangeIndicationEnum>> getter,
       Consumer<HomekitCharacteristicChangeCallback> subscriber,
       Runnable unsubscriber) {
     super(
-        "00000071-0000-1000-8000-0026BB765291",
-        "Occupancy Detected",
-        OccupancyDetectedEnum.values(),
+        "000000AC-0000-1000-8000-0026BB765291",
+        "filter change indication",
+        FilterChangeIndicationEnum.values(),
         Optional.of(getter),
         Optional.empty(),
         Optional.of(subscriber),