+++ src/main/java/io/github/hapjava/characteristics/impl/humidifier/TargetHumidifierDehumidifierStateCharacteristic.java
@@ -15,6 +15,7 @@ public class TargetHumidifierDehumidifierStateCharacteristic
     implements EventableCharacteristic {
 
   public TargetHumidifierDehumidifierStateCharacteristic(
+      TargetHumidifierDehumidifierStateEnum[] validValues,
       Supplier<CompletableFuture<TargetHumidifierDehumidifierStateEnum>> getter,
       ExceptionalConsumer<TargetHumidifierDehumidifierStateEnum> setter,
       Consumer<HomekitCharacteristicChangeCallback> subscriber,
@@ -22,18 +23,10 @@ public class TargetHumidifierDehumidifierStateCharacteristic
     super(
         "000000B4-0000-1000-8000-0026BB765291",
         "target humidifier state",
-        2,
+        validValues,
         Optional.of(getter),
         Optional.of(setter),
         Optional.of(subscriber),
         Optional.of(unsubscriber));
-  }
-
-  @Override
-  protected void setValue(Integer value) throws Exception {
-    if (!setter.isPresent()) {
-      return;
-    }
-    setter.get().accept(TargetHumidifierDehumidifierStateEnum.fromCode(value));
   }
 }