+++ src/main/java/io/github/hapjava/services/impl/AbstractServiceImpl.java
@@ -5,13 +5,11 @@ import io.github.hapjava.services.Service;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 abstract class AbstractServiceImpl implements Service {
-  private final Logger logger = LoggerFactory.getLogger(this.getClass());
   private final String type;
   private final List<Characteristic> characteristics = new LinkedList<>();
+  private final List<Service> linkedServices = new LinkedList<>();
 
   /** @param type unique UUID of the service according to HAP specification. */
   public AbstractServiceImpl(String type) {
@@ -28,7 +26,17 @@ abstract class AbstractServiceImpl implements Service {
     return type;
   }
 
+  @Override
+  public List<Service> getLinkedServices() {
+    return Collections.unmodifiableList(linkedServices);
+  }
+
   public void addCharacteristic(Characteristic characteristic) {
     this.characteristics.add(characteristic);
+  }
+
+  @Override
+  public void addLinkedService(Service service) {
+    this.linkedServices.add(service);
   }
 }